refactor: migrate API routes to Next.js 15.1.5 app directory structure
